Refactor CafienneGateway ==> CaseEngineGateway
Also changed the routing logic to use the actual actor class of the ModelCommand

Also merged CaseMessageRouter and LocalRouter classes to avoid switching between those classes to understand what is happening
(and the sharding version is yet to be developed)
